ref: add NotFoundException to be caught in the exceptionHandler

+++ src/main/java/br/edu/ifpe/easy_football_management_backend/features/championshipsteams/command/ChampionshipsTeamsCommandHandller.java
@@ -1,6 +1,7 @@
 package br.edu.ifpe.easy_football_management_backend.features.championshipsteams.command;
 
 import br.edu.ifpe.easy_football_management_backend.application.commom.exceptions.BusinessException;
+import br.edu.ifpe.easy_football_management_backend.application.commom.exceptions.NotFoundException;
 import br.edu.ifpe.easy_football_management_backend.domain.entity.*;
 import br.edu.ifpe.easy_football_management_backend.features.championshipsteams.ChampionshipsTeamsDTO;
 import br.edu.ifpe.easy_football_management_backend.features.championshipsteams.ChampionshipsTeamsMapper;
@@ -30,7 +31,9 @@ public class ChampionshipsTeamsCommandHandller {
     public ChampionshipsTeamsCommandHandller(
             ChampionshipsTeamsRepository championshipsTeamsRepository,
             TeamRepository teamRepository,
-            TokenService tokenService, ResultRepository resultRepository, RedissonClient redissonClient,
+            TokenService tokenService,
+            ResultRepository resultRepository,
+            RedissonClient redissonClient,
             ChampionshipsRepository championshipsRepository,
             ApplicationEventPublisher eventPublisher) {
         this.championshipsTeamsRepository = championshipsTeamsRepository;
@@ -44,14 +47,14 @@ public class ChampionshipsTeamsCommandHandller {
 
     @Transactional
     public ChampionshipsTeams create(String authHeader, ChampionshipsTeamsDTO championshipsTeamsDTO) {
-        String keyLock = "ChampionshipsTeamsCommandHandller.create." + championshipsTeamsDTO.championshipsId() + "." + championshipsTeamsDTO.teamId();
+        String keyLock = "ChampionshipsTeamsCommandHandller.create.";
         var lock = redissonClient.getLock(keyLock);
         lock.lock();
         UUID userId = UUID.fromString(tokenService.extractID(authHeader));
         Optional<UUID> optionalTeamId = teamRepository.findFirstTeamIdByUserId(userId);
         Championships championships = championshipsRepository
                 .findById(championshipsTeamsDTO.championshipsId())
-                .orElseThrow(() -> new BusinessException("Champions not found"));
+                .orElseThrow(() -> new NotFoundException("Champions not found"));
         if (optionalTeamId.isEmpty() || !optionalTeamId.get().equals(championshipsTeamsDTO.teamId())) {
             throw new BusinessException("Team does not belong to the user");
         }
@@ -59,7 +62,9 @@ public class ChampionshipsTeamsCommandHandller {
         if (championshipsTeamsRepository.existsByTeamId(championshipsTeamsDTO.teamId())) {
             throw new BusinessException("Team already exists");
         }
-        ;
+        if (countTeamInChampionship >= championships.getQuantityTeams()) {
+            throw new BusinessException("Championship already has the maximum number of teams");
+        }
         ChampionshipsTeams entity = mapper.toEntity(championshipsTeamsDTO, teamRepository.findById(championshipsTeamsDTO.teamId())
                 .orElseThrow(() -> new BusinessException("Teams not found")), championships);
         countTeamInChampionship++;
@@ -74,13 +79,16 @@ public class ChampionshipsTeamsCommandHandller {
     public void deleteTournamentsTeams(String authHeader, UUID ChampionshipsId) {
         UUID ID = UUID.fromString(tokenService.extractID(authHeader));
         Optional<UUID> optionalTeamId = teamRepository.findFirstTeamIdByUserId(ID);
-        UUID teamId = optionalTeamId.orElseThrow(() -> new BusinessException("Team not found"));
-        String keyLock = "ChampionshipsTeamsCommandHandller.create." + ChampionshipsId + "." + teamId;
+        UUID teamId = optionalTeamId.orElseThrow(() -> new NotFoundException("Team not found"));
+        String keyLock = "ChampionshipsTeamsCommandHandller.create.";
         var lock = redissonClient.getLock(keyLock);
         lock.lock();
         Integer countTeamInChampionship = championshipsTeamsRepository.countByTeamContains(ChampionshipsId);
-        var championship = championshipsRepository.findById(ChampionshipsId).orElseThrow(() -> new BusinessException("Championship not found"));
+        var championship = championshipsRepository.findById(ChampionshipsId).orElseThrow(() -> new NotFoundException("Championship not found"));
         boolean isExist = championshipsTeamsRepository.existsByTeamId(teamId);
+        if (countTeamInChampionship.equals(0)) {
+            throw new BusinessException("Championship does not have any teams");
+        }
         if (isExist && countTeamInChampionship.equals(championship.getQuantityTeams())) {
             resultRepository.deleteByChampionshipsId(ChampionshipsId);
         }
